re-arranged, have not re-connected variables yet
