+++ sizing/sizing_comp.py
